LayoutTests/platform/mac-wk1/http/tests/security/http-0.9/image-on-HTTP-0.9-page-blocked-expected.txt

CONSOLE MESSAGE: Sandboxing 'http://127.0.0.1:8000/security/http-0.9/resources/nph-image-on-HTTP-0.9-page-blocked.pl' because it is using HTTP/0.9.


--------
Frame: '<!--frame1-->'
--------

http://127.0.0.1:8000/security/http-0.9/image-on-HTTP-0.9-page-blocked.html - willSendRequest <NSURLRequest URL http://127.0.0.1:8000/security/http-0.9/image-on-HTTP-0.9-page-blocked.html, main document URL http://127.0.0.1:8000/security/http-0.9/image-on-HTTP-0.9-page-blocked.html, http method GET> redirectResponse (null)
http://127.0.0.1:8000/security/http-0.9/image-on-HTTP-0.9-page-blocked.html - didReceiveResponse <NSURLResponse http://127.0.0.1:8000/security/http-0.9/image-on-HTTP-0.9-page-blocked.html, http status code 200>
http://127.0.0.1:8000/security/http-0.9/resources/nph-image-on-HTTP-0.9-page-blocked.pl - willSendRequest <NSURLRequest URL http://127.0.0.1:8000/security/http-0.9/resources/nph-image-on-HTTP-0.9-page-blocked.pl, main document URL http://127.0.0.1:8000/security/http-0.9/image-on-HTTP-0.9-page-blocked.html, http method GET> redirectResponse (null)
http://127.0.0.1:8000/security/http-0.9/image-on-HTTP-0.9-page-blocked.html - didFinishLoading
http://127.0.0.1:8000/security/http-0.9/resources/nph-image-on-HTTP-0.9-page-blocked.pl - didReceiveResponse <NSURLResponse http://127.0.0.1:8000/security/http-0.9/resources/nph-image-on-HTTP-0.9-page-blocked.pl, http status code 200>
http://127.0.0.1:8080/security/http-0.9/resources/nph-image.pl - willSendRequest <NSURLRequest URL http://127.0.0.1:8080/security/http-0.9/resources/nph-image.pl, main document URL http://127.0.0.1:8000/security/http-0.9/image-on-HTTP-0.9-page-blocked.html, http method GET> redirectResponse (null)
http://127.0.0.1:8000/security/http-0.9/resources/nph-image-on-HTTP-0.9-page-blocked.pl - didFinishLoading
http://127.0.0.1:8080/security/http-0.9/resources/nph-image.pl - didFailLoadingWithError: <NSError domain , code 0, failing URL "http://127.0.0.1:8080/security/http-0.9/resources/nph-image.pl">
